Bank Management System In C++
This is a very common project and can make you a pro in C++, all you have to do is

Open an account
Show account details
Deposit
Withdraw
Search and exit.
This project will let you understand the fundamentals of banking apps. In a nutshell, you can say that this primary C++ project is going to familiarize you with data adjustment, how to add, modify or delete any filed record. Hence, it is a very crucial project that can enhance your mastery of C++.